Move API fetch related classes do fetcher package
