+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/fetcher/ApiResult.kt
@@ -1,0 +1,6 @@
+package com.gustavo.competitiveprogrammingapp.cfApi.fetcher
+
+class ApiResult<T> {
+    val status: String? = null
+    val result: T? = null
+}
