+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/fetcher/UrlCacheRepository.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi
+package com.gustavo.competitiveprogrammingapp.cfApi.fetcher
 
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
